+++ src/com/intellij/idea/plugin/hybris/system/type/model/CustomProperty.java
@@ -1,6 +1,6 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
- * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -37,6 +37,8 @@ import org.jetbrains.annotations.NotNull;
  */
 public interface CustomProperty extends DomElement {
 
+    String NAME = "name";
+
     /**
      * Returns the value of the name child.
      * <pre>
@@ -47,7 +49,7 @@ public interface CustomProperty extends DomElement {
      * @return the value of the name child.
      */
     @NotNull
-    @com.intellij.util.xml.Attribute("name")
+    @com.intellij.util.xml.Attribute(NAME)
     @Required
     GenericAttributeValue<String> getName();